Break BlockstoreError into separate file
